feat: convert api to TS & move `getRota` to an api file
