--- src/services/api.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-const apiUrl = import.meta.env.VITE_PROXY_URL;
-
-const api = axios.create({
-  baseURL: apiUrl
-});
-
-export default api;
